+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - Removed dependency constraint of `dagster-pandas` on `pandas`. You can now include any version of pandas. (https://github.com/dagster-io/dagster/issues/3350)
 - Removed dependency on `requests` in `dagster`. Now only `dagit` depends on `requests.`
+- Removed dependency on `pyrsistent` in in `dagster`
 
 ## 0.11.4